Preliminary work for OPENIG-66 - Add support for Groovy based scripting
"If it looks like a duck, swims like a duck, and quacks like a duck, then it probably is a duck."

But Ducks are fat and we don't want lots of excess fat in our code base so, since we are not duck collectors, let's go shoot some duck. :-)

Seriously: OpenIG uses duck-typing to provide map-like views of various objects. JSR 223 implementations, such as for Groovy, provide automatic support for bean-like objects and maps. In particular, bean getters/setters and map key/value pairs are automatically mapped to properties in Groovy scripts. This makes scripting really easy and developer friendly. Unfortunately, the support does not extend to our duck-typed objects, which are neither beans or maps. I was unable to see why duck-typing is needed in OpenIG, other than to avoid implementing a couple of awkward map view methods (e.g. entrySet), but they bring their own complexity. For example, many engineers find the duck-type support hard to understand and maintain.