Provide the ability to evaluate JsonValue containing expressions 'in place'.
The previously introduced evaluate() method was a step in the right
direction, allowing resolution of static (no references to ${exchange})
string JsonValue, returning a simple resolved String.
While it works great, it's a shame that we don't benefit anymore of the
JsonValue API, in particular the conversion methods (asURL(), ...).

This fix uses a JsonTransformer to actually resolve any String wrapped
JsonValue and creates a deep copy of the given JsonValue.