Expose new security heap objects (KeyStore, KeyManager and TrustManager)
KeyStore and related security objects are being more and more central
to OpenIG:
 * Used by HttpClient
 * Needed for encryption / decryption

This patch is a first step to provide a way to configure theses as
usual heap objects (through JSON).

KeyStore heap object is an abstraction around the keystore file (either JKS,
PKCS12 or other depending on the platform capabilities).
Specifying a password or not depends on he usage:
 * If KeyStore will be used to read private credentials (through a
   KeyManager for example), a password is required
 * If not (for a TrustManager), no password is required

KeyManager and TrustManager are using a KeyStore heap object and
are used within SSL mechanism (SSLContext) to provide a view on the KeyStore:
 * KeyManager when private credentials access is required.
 * TrustManager when only public information access is required

Issues: OPENIG-295
Reviews: CR-4441