CR-4573 OPENIG-275: Document what is needed for load balancing and failover
If you can store all state on the user-agent,
for example by using the JwtSession implementation,
then perhaps OpenIG can be stateless enough
that there is no need to do anything special when load balancing.

If some of the state is stored on the server,
then you need to configure the load balancer for session stickiness
and to configure the container for session replication.

Neither the load balancer configuration
nor the container configuration
are specific to OpenIG,
so this patch explains what needs doing
and points to the documentation
for supported containers Apache Tomcat & Jetty.