Supports multiple declaration resolution.
This case can happen when a component tries to resolve twice the same
object declaration.

This is unlikely to happen without any decorator support because all the
heap interactions are going on in one place and it's easy to refactor the
code to only resolve a declaration once.

But when decorators comes into play, it's no more easy to share resolved
objects references around.