OPENIG-299 (CR-4904) Add capture decoration for Filter/Handler
This decorator can be applied on any Filters or Handlers whose
input/outputs needs to be observed.

This decorator prints the message's values in the observed
component's logSink or in its own configurable logSink (if one
was specified).

Multiple observation points can be specified depending on what you're
interested in.

Observation points for Filter:
                            ---
          -- (request) ->  |   | -- (filtered-request) ->
                           | F |
 <- (filtered-response) -- |   | <- (response) --
                            ---

Observation points for Handler:
                   ---
 -- (request) ->  |   |
                  | H |
 <- (response) -- |   |
                   ---