OPENIG-339 (CR-4964) Provide more log information when route building fails
Exception's messages does not have to include the message of their parent anymore.
When a route failed to load for any reason, all the chained exception's
messages are logged as errors up to the root cause, and the full stack
trace is logged at the DEBUG level.

Notice that the LogSink you have to configure is the one of the Router itself (in
the JSON configuration where the Router handler is declared), configuring the LogSing
of the failed route configuration will not help.