OPENIG-301 Add an option to dump the exchange as part of the capture
The CaptureDecorator can now be configured to additionally capture the content
of the Exchange being intercepted.

If specified, it will print the full content of the Exchange instance, excluding
the `response` and the `request` since they should be captured by another mean.
The output contains a deep view of the exchange obtained through introspection
of the class structure (this is done by the Groovy JsonOutput class).
It appears that this class doesn't support cycles very well, so, depending on
the content of the exchange, it may throw exceptions.
By default we're doing our best to remove known cycles.

By default (in the system provided `capture` decorator), this option is disabled
since it adds a lot of cruft in the logs.