OPENIG-361 Provides tools for decorators to avoid StackOverFlowErrors at creation time
A `StackOverFlowError` can be thrown during the Heap init when the configuration file declare
global decorators that have a heap object dependency.

When the globally enabled decorator is first created, it tries to resolve a dependency
from the heap, the heap then tries to decorate that instance, looking for the globally
declared decorator, that is not yet available since it has not finished its initialization,
so the heap think the decorator instance was not created yet and triggers another
decorator instance creation, that will itself try to resolve the dependency, looping
again and again ad nauseam.

The framework cannot provide any guards against that problem right now, the decorators
implementers have to care of this on their own and carefully craft their decorators
to avoid that problem.

The framework can only provide some limited level of support to help developers not
hitting that issue.

Introduced a new `DecoratorHeaplet` abstract class that does not resolve automatically heap
objects at creation time (no `LogSink` and `TemporaryStorage` resolution, as opposed to
`GenericHeaplet` behaviour).

Introduced a `LazyReference<T>` that encapsulate the resolution logic to allow easy
heap object resolution delaying.

Decorator implementation are encouraged to use theses 2 classes (having their `Heaplet`
extending `DecoratorHeaplet` instead of `GenericHeaplet`) and using `LazyReference` when a
heap object dependency is un-avoidable.

Moved `CaptureDecorator` to use the `LazyReference` and updated existing decorator's
heaplet to extend `DecoratorHeaplet`.

Updated javadoc to make that clear for Decorator implementers.