(CR-5558) Add an error message that can be displayed by the default router
Only a `404` status code was returned when something goes wrong, nothing was
actually printed on screen if you accessed OpenIG from a browser.

This change add a JSON content (still machine parseable) with a descriptive
error message that is displayed by browser, making the client aware that
something was wrong for any reason.

This change was triggered with fix for OPENIG-411, that had the side effect of
not propagating route's exception to the user anymore