OPENIG-315 (CR-5787) Fix non-empty messages being considered as empty
The `Entity.isEmpty()` method relied on `InputStream.available()` returning
something different of `0` (zero) to detect if the entity was empty or not.
This technic is not reliable enough since the javadoc clearly states that
this return the number of bytes that can be read **without blocking**, that
means that the result of this method depends on the network (and to some
extension to the web container), not entirely on the message.

The idea here is to now check if the wrapped stream is the `EMPTY_STREAM` instance
to detect if the entity is empty or not.
This is better than the old solution because we don't rely anymore on an external
`InputStream` implementation, but that will not detect a user provided stream with
no data inside. This is why `isEmpty()` has been renamed to `mayContainData()`.
There is also a new `setEmpty()` method to mark the entity as empty (simply assign
`EMPTY_STREAM` to the wrapped stram field).