OPENIG-315 Complete the fix for empty messages
The previous commit introduced a regression, especially for GET messages
being sent with an empty entity where they should have no entity at all.

They were being falsely detected as messages with content because we were
comparing `EMPTY_STREAM` with `Entity.head` (the branched content) instead
of the `Entity.trunk` (`head` being re-created every time OpenIG tries to read
the content, so in the `HttpBasicAuthFilter` for example, we push the Entity,
then delegates to the next handler in chain, so when message is sent, the
`head` is != from the `trunk` and cannot be == to `EMPTY_STREAM`)

I could not reliably reproduce that in a unit test with a real HTTP Server
because the failure also need the HTTPClient to re-use the same connection
for 2 consecutive messages. Thus, I only added a small `Entity` unit test
to make sure that `Entity.mayContainData()` behave correct even when the
entity is pushed/popped.