OPENIG-180 (CR-5830) Provide a baseURI decorator As the URI rebasing is done at different places in the code we'd like to have a BaseUriHandler/Filter in order to factor out the code.
Following the same scheme as the "Timer" decorator, the "baseURI"
decorator is created by default in the Gateway Servlet.
(Named "baseUri" and created at startup time in the top-level heap.)

* GatewayServlet.class, Route.class
The creation of the "baseUri" decorator means the attribute
class 'baseURI' is no longer needed as the URI rebasing is now
directly done by the decorator. In the other hand,
the heap initialization performed within both class constructors,
contained a list of reservedFieldNames where the 'baseURI'was present.
It has been removed from there as it is now a global decorator.

* RouteTest.java
Removed unit test 'testRouteIsRebasingTheRequestUri'(Duplicated
in the RouteBuilder test, and the RouteBuilder has the responsability
to apply decorators).