OPENIG-443 (CR-6009) Fix erroneous Response emptying
`HttpClient` is closing the `exchange.response` object before executing the request.
When the `OAuth2ClientFilter` populates the `user_info` attribute (requesting user's
profile on the AS), or communicate with the AS, it reuses the `Exchange`, change
the request object, but keep the response reference.
So, when the modified exchange goes through `HttpClient`, any existing response
gets emptied.

This fix simply nullify the `exchange.response` reference *before* sending that
to the AS. That prevents `HttpClient` to cleanup the response too eagerly.