CR-6746 OPENIG-490 Allow simultaneous use of all example routes
This patch updates the examples in the documentation
so that they are all routes and can be used at the same time.

Some examples still require additional configuration.

This patch also updates the sample doc server
for use with the revised examples.

This patch also adds support for authentication with headers,
as shown in the following example:

$ curl \
 --request POST \
 --header "username: bjensen" \
 --header "password: hifalutin" \
 http://localhost:8081/

This capability is not yet used in the docs,
but in review we found it could be useful for QA.