Fix compilation error due to changeof the Status.
Since we introduce the class Status in forgerock-http-framework, some code's
modifications have to be done :
- To get the code of the status we now have to do response.getStatus().getCode()
into of response.getStatus(). (in the Groovy script or EL expression that means
we'll have to write response.status.code.
- Furthermore, the attribute reason in Response does not exist any more, so it
is not possible anymore to assign a custom reason phrase, as it is set to
standardized reason phrase through the constants defined in the class Status.
- Any status code has to be in the range [100,1000[