OPENIG-528 Remove `ExceptionFilter`
This change is a follow-up on CHF-34 change (`Promise<Response, NeverThrowsException>`).
There is no more any exception to use in the `ExceptionFilter`.
In other words, `ExceptionFilter` is useless in this new context.

User of `ExceptionFilter` are encouraged to use a `SwitchFilter` instead
(leveraging the response status in the condition expression):

```
{
    "name": string,
    "type": "SwitchFilter",
    "config": {
        "onResponse": [
            {
                "condition": "exchange.response.status.code == 500",
                "handler": Handler reference,
            },
            ...
        ]
    }
}
```

Documentation has been cleaned out too.