OPENIG-543 (CR-7114) Update HttpClient API to support non-blocking calls
`Handler` and `Filter` are now returning `Promises` (thus are non blocking from
an API point of view).
The CHF `Client` SPI also supports non-blocking calls.

But all elements of the processing chain need to support this in order to really
gain something out of the `Promises`.

The `HttpClient` was not offering such support, so the OpenIG `ClientHandler` could
not take advantage of it.