OPENIG-545 (CR-7225) Fixes NPE in SamlFederationHandler
The NPE was caused because since r1010, there is no more any reference to the
`HttpServletRequest` and `HttpServletResponse` objects provided by the web
container (was done to prevent context serialization issues).

This new version of the fix now ensure that we place the servlet objects at the
right place (directly in the `Exchange`, where the message capture code is
intentionally excluding them from the capture) so that the `SamlFederationHandler`
can still get theses references.

We're also removing the content of the `httpContext.getAttributes()` Map so that
the message capture mechanism (when capturing the contexts) will not try to
serialize theses objects (because we know that it's failing right now).