CR-7428 OPENIG-522 OPENIG-463 Provides OpenID Connect Discovery and Dynamic Client Registration support
It introduces new objects and new filters for the OpenIG configuration
that allow discovery process and client registration to the AS.
The process is launched by the OAuth2ClientFilter, by a request on an URL
containing the 'discovery' pattern.

(Note: This is the first iteration for those issues. The code should change
soon). See JIRA issue for further details.