OPENIG-469 Drop ExtensibleFieldMap
The EntensibleFieldMap is removed. As the Exchange was using it to
store its misc attributes, commonly used in expressions
like ${exchange.var}, a Map 'attributes' field was added to
the Exchange. Now, with the attributes field, accessing misc var in
expression should be: ${exchange.attributes.var}
Moreover, data encapsulation was increased on the Exchange class.