OPENIG-604 Doc changes for OIDC discovery and registration
This patch lets the sample doc app respond to WebFinger requests.
It also gets the tests working again for the sample doc app.

The intent with this patch is to allow the sample app to assist
with OpenID Connect discovery in an doc example to show
discovery and dynamic registration.