Fix inter-leaved log entries on the console
You can have multiple `ConsoleLogSink` instance in your configuration, but
at the end, they all print in the same `PrintStream` (`System.out`, or
`System.err`). Having a `synchronised` block only on `this` was not sufficient
to guarantee that entries were not interleaved.

Using a `static` lock Object make the lock work for all `ConsoleLogSink` instances.