OPENIG-694 Introduce SsoTokenFilter
This filter decorate the filtered request object with an OpenAM SSO Token
that it obtains from the `/json/authenticate` endpoint.

The obtained token is stored in session to avoid un-necessary calls to OpenAM.

The filtered request in retried only once if the response is `FORBIDDEN`
(that means that the provided token has expired, or was revoked). If the response
fails one more time, this filter just let it flow back to the caller.