OPENIG-794 Fix typos in DispatchHandler.groovy example script
This patch fixes typos that were causing the example in the docs to fail.

It seems like the bottom part of the file was aligned with the doc
where I think we moved to `/mylogin` to avoid a clash with `/login` elsewhere,
but the top part of the file and the tests were not changed at that time.