OPENIG-809 Apply workaround for concurrent issues on blocking calls
This function is here to fix a concurrency issue where a caller thread is
blocking a promise and is resumed before all of the ResultHandlers and
Function of the blocked promise have been invoked.

That may lead to concurrent consumption of BranchingInputStream that is
a notably **not** thread safe object.