OPENIG-712 Better implementation for supportedDomains
The list of client registrations are now listed in the OAuth2ClientFilter
they depend on. This eases the configuration and retrieving supported domains
is more reliable than before (the tip mentioned in this issue about
client registration name no longer works).

A new class 'ClientRegistrationRepository' was added to store the different
client registrations instead of relying on the heap, solving OPENIG-581.

Thanks to Laurent and Guillaume for their help :)