OPENIG-825 Use defensive copy of Request
CHF-114 provides defensive copies for requests and responses.
In OPENIG, HttpBasicAuthFilter and PasswordReplayFilterHeaplet classes need to process the request
without modifying the original. Indeed, the original request is reused in these classes, after these processes.
The new 'RequestCopyFilter', used to create defensive copies of the requests, will be used in both classes
to simplify the actual code and to ensure that the original request is not altered by the upstream part.