OPENIG-933 Handle token expiration in the OAuth2ClientFilter
As we saw, when the access token issued from Authorization Server has
expired, then new requests to the Authorization Server fails
(OAuth2ErrorException: error="invalid_request"). Previous [code
cleanup](087f73378c7) demonstrates that the code responsible to refresh
the token when the token expires was never called.

To handle token expiration, we have to use the refresh token to create a
new access token to access the protected resource.