OPENIG-953 Implements new HeapClientRegistrationRepository
The ClientRegistrationRepository was not compatible with OPENIG 4.0 and
it wasn't using the heap anymore so the HeapClientRegistrationRepository
class was created to fill the gap. This patch insure compatibility between
version 4.0 and 4.5.

Indeed, the OAuth2ClientFilter is searching given ClientRegistration in the heap
even if it is not declared in the "registrations" attributes.