OPENIG-1115 Serve OpenIG UI in /openig/console
The `openig-ui.jar` content is now unpacked in the `${openig.base}/tmp/openig-ui` directory.

When unpacking, we try to be smart and not unpacking again resources that
have not changed.

We try to keep local modification of unpackaged resources: they will not be
overridden if the local resource are newer (but they will be overridden
if the jar resources are newer).