This filter is responsible to force OpenAM-based authentication. In the absence (or invalidity) of the ssoToken in the request cookie header, it redirects the User-Agent on the appropriate OpenAM Login page.
In case of a valid token and a valid session, the request is forwarded
to the next filter in chain. The session is then completed with an
additional SsoTokenContext, used to store the json result
from the validation endpoint. The ssoToken is added here
(ex: ${contexts.ssoToken.value}) with additional info
(${contexts.ssoToken.info}).